requirements.txt
pandas~=2.2.3
requests~=2.32.3
jmespath~=1.0.1
keyring~=25.6.0
python-dotenv~=1.0.1
loguru~=0.7.3